Refact: Renaming RoomsList to Rooms
