+++ PlanningPoker.Client/Pages/Rooms/Rooms.razor
@@ -1,5 +1,5 @@
 @page "/Rooms"
-@using PlanningPoker.Client.Pages.RoomsList.Components
+@using PlanningPoker.Client.Pages.Rooms.Components
 
 @if (_roomCollection is null)
 {
