+++ PlanningPoker.Client/Pages/Rooms/Rooms.razor.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Components;
 using PlanningPoker.Client.Services.Rooms;
 using PlanningPoker.Shared.Dtos;
 
-namespace PlanningPoker.Client.Pages.RoomsList;
+namespace PlanningPoker.Client.Pages.Rooms;
 
 public partial class Rooms
 {
